rzg3: linux-renesas: Make WIFI firmware download as conditional option
Currently, the Intel Wifi firmware of AX200 is forcely downloaded
no matter AX200 is enabled or not. This patch make the download as
conditional option and depend on AX200 selection status.

+++ meta-rzg3/recipes-kernel/linux/linux-renesas_5.10.bbappend
@@ -14,7 +14,7 @@ SRC_URI_remove = " \
 IWLWIFI_FIRMWARE = "https://git.kernel.org/pub/scm/linux/kernel/git/iwlwifi/linux-firmware.git/plain/iwlwifi-cc-a0-46.ucode;md5sum=babe453e0bc18ec93768ec6f002d8229;downloadfilename=iwlwifi-cc-a0-46.ucode"
 
 SRC_URI_append = " \
-	${IWLWIFI_FIRMWARE} \
+	${@bb.utils.contains('MACHINE_FEATURES', 'ax200-wifi', '${IWLWIFI_FIRMWARE}', '',d)} \
 	${@bb.utils.contains('MACHINE_FEATURES', 'ax200-wifi', 'file://ax200-wifi.cfg', '',d)} \
 	${@bb.utils.contains('MACHINE_FEATURES', 'atheros-ar9287-wifi', 'file://atheros-ar9287-wifi.cfg', '',d)} \
 	${@bb.utils.contains('MACHINE_FEATURES', 'rtl8169-firmware', 'file://rtl8169-firmware.cfg', '',d)} \
@@ -23,7 +23,9 @@ SRC_URI_append = " \
 
 BUILTIN_FIRMWARE_DIR = "${STAGING_KERNEL_DIR}/drivers/base/firmware_loader/builtin"
 do_download_firmware () {
-	install -m 755 ${WORKDIR}/iwlwifi-cc-a0-46.ucode ${BUILTIN_FIRMWARE_DIR}
+	if [ -f ${WORKDIR}/iwlwifi-cc-a0-46.ucode ]; then
+		install -m 755 ${WORKDIR}/iwlwifi-cc-a0-46.ucode ${BUILTIN_FIRMWARE_DIR}
+	fi
 }
 
 addtask do_download_firmware after do_configure before do_compile
